Abstracted note highlight config to HighlightConfig component
