+++ src/components/Guitar/Guitar.scss
@@ -16,9 +16,3 @@
 .guitar {
   width: 100%;
 }
-
-.colour-block {
-  display: inline-block;
-  width: 1rem;
-  height: 1rem;
-}
