+++ src/components/Guitar/Guitar.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
   import Config from './subcomponents/Config/Config.svelte';
+  import HighlightConfig from './subcomponents/HighlightConfig/HighlightConfig.svelte';
   import String from './subcomponents/String/String.svelte';
   import type { SelectedNote } from './types/SelectedNote';
 
@@ -14,21 +15,6 @@
   let highlightedNotes: SelectedNote[] = [];
 
   let guitarElement: HTMLElement;
-
-  $: selectedNoteIsHighlighted = highlightedNotes.some(({ value }) => value === selectedNote?.value);
-
-  function highlightNote(note: SelectedNote) {
-    const noteIndex = highlightedNotes.findIndex(highlightedNote => highlightedNote.value === note.value);
-    const noteIsAlreadyHighlighted = noteIndex > -1;
-    if (!noteIsAlreadyHighlighted) {
-      highlightedNotes = [...highlightedNotes, note];
-    } else {
-      highlightedNotes = [
-        ...highlightedNotes.slice(0, noteIndex),
-        ...highlightedNotes.slice(noteIndex + 1)
-      ];
-    }
-  }
 
   function handleWindowClick({ target }: MouseEvent) {
     assertEventTargetIsNode(target);
@@ -66,34 +52,10 @@
     {/each}
   </div>
   {#if selectedNote}
-    <div>
-      <h1>
-        {selectedNote.value}
-      </h1>
-      <div>
-        <label>
-          <span>Colour</span>
-          <input type="color" bind:value={selectedNote.color} />
-        </label>
-        <div
-          class="colour-block"
-          style="background-color: {selectedNote.color};"
-        />
-        <span>
-          {selectedNote.color}
-        </span>
-      </div>
-      <div>
-        <label>
-          <span>Name</span>
-          <input type="text" bind:value={selectedNote.name} />
-        </label>
-      </div>
-      <button on:click={() => selectedNote && highlightNote(selectedNote)}>
-        {selectedNoteIsHighlighted ? 'Remove' : 'Add'}
-      </button>
-      <hr>
-    </div>
+    <HighlightConfig
+      {selectedNote}
+      bind:highlightedNotes={highlightedNotes}
+    />
   {/if}
   <button
     on:click={() => displayConfig = !displayConfig}
