+++ src/components/Guitar/subcomponents/HighlightConfig/HighlightConfig.scss
@@ -1,0 +1,5 @@
+.colour-block {
+  display: inline-block;
+  width: 1rem;
+  height: 1rem;
+}
